Adicionando atividades feitas da lista1 lista2 e avaliaçãol1

+++ Lista2/ex01-area-do-circulo.c
@@ -2,9 +2,12 @@
 
 int main()
 {
-    float area, n, raio;
-    scanf("%f.2", &raio);
+    double area, n, raio;
+
+    scanf("%lf.2", &raio);
+
     area = ((raio * raio) * 3.14159);
+    
     printf("A=%.4f\n", area);
 
     return 0;
